Add adhan sound to prayer notification
This commit introduces the ability to play an adhan sound when a prayer time notification is shown.

A new setting "Play Adhan Sound" has been added to the plugin configuration, allowing users to enable or disable this feature.

The `FocusModeDialog` now checks this setting and, if enabled, plays an embedded "adhan.mp3" sound when the dialog is displayed.